+++ client/actions/actionTypes.js
@@ -14,3 +14,4 @@ export const RECEIVED_FR = 'RECEIVED_FR';
 export const SENT_FR = 'SENT_FR';
 export const CONFIRMED_FR = 'CONFIRMED_FR';
 export const CONFIRMED_FR2 = 'CONFIRMED_FR2';
+export const YES_COMPETITIONS = 'YES_COMPETITIONS';
