+++ client/components/index.js
@@ -8,11 +8,10 @@ export SearchResults from './searchResults';
 export UserProfile from './userProfile';
 export Repos from './repos';
 export RepoProfile from './repoProfile';
-export OrgProfile from './orgProfile';
 export CumulativeChart from './../chartmakers/cumulativeChart';
 export DailyChart from './../chartmakers/dailyChart';
 export Organizations from './organizations';
 export UserAchievements from './UserAchievements';
 export ChooseWeapon from './chooseWeapon';
 export ChooseSecondWeapon from './chooseSecondWeapon';
-export CompetitorsMiniView from './CompetitorsMiniView';
+export CompetitorsMiniView from './competitorCards/competitorsMiniView';
