+++ client/index.js
@@ -25,7 +25,6 @@ render(
         <Route path="/:username/profile" component={component.UserProfile} />
         <Route path="/:username/achievements" component={component.UserAchievements} />
         <Route path="/:username/repos" component={component.Repos} />
-        <Route path="/:orgname/orgs" component={component.OrgProfile} />
         <Route path="/compete/choose-repo/:username" component={component.ChooseWeapon} />
         <Route path="/compete/choose-second-repo/:userid" component={component.ChooseSecondWeapon} />
       </Route>
